+++ CHANGELOG.md
@@ -30,6 +30,7 @@
    * UPDATED: submodules which had new releases, unless it was a major version change [#4231](https://github.com/valhalla/valhalla/pull/4231)
    * ADDED: the workflow to find landmarks in a graph tile, associate them with nearby edges, and update the graph tile to store the associations [#4278](https://github.com/valhalla/valhalla/pull/4278)
    * ADDED: update maneuver generation to add nearby landmarks to maneuvers as direction support [#4293](https://github.com/valhalla/valhalla/pull/4293)
+   * ADDED: update maneuver types and narrative generation to support the use of landmarks in phrases, specially for `turn`, `bear`, `sharp`, and `uturn` instructions [#4301](https://github.com/valhalla/valhalla/pull/4301)
 
 ## Release Date: 2023-05-11 Valhalla 3.4.0
 * **Removed**